requirements.txt
python-telegram-bot==22.6.0
flask